Upgrade AppDelegate code
