+++ Smogler/LocationManager+iOS.swift
@@ -1,0 +1,9 @@
+//
+//  LocationManager+iOS.swift
+//  Smogler
+//
+//  Created by Dominik Bucher on 26/05/2018.
+//  Copyright © 2018 vasarhelyia. All rights reserved.
+//
+
+import Foundation
